modify Valdation move to function js
