+++ src/pages/Main/Login/LoginInput.js
@@ -1,8 +1,10 @@
 import React, { useState } from 'react';
+import { InputValueSignup } from '../../../Function';
 import Modal from '../../../components/Modal/Modal';
 import './LoginInput.scss';
 
 const LoginInput = ({ isLoginMode }) => {
+  const [isShowModal, setIsShowModal] = useState(false);
   const [inputValues, setInputValues] = useState({
     username: '',
     password: '',
@@ -11,17 +13,6 @@ const LoginInput = ({ isLoginMode }) => {
     name: '',
     address: '',
   });
-
-  const [isShowModal, setIsShowModal] = useState(false);
-
-  const togleModal = () => {
-    setIsShowModal(togleModal => !togleModal);
-  };
-
-  const handleInput = event => {
-    const { name, value } = event.target;
-    setInputValues({ ...inputValues, [name]: value });
-  };
 
   const loginSignUp = (e, urlSignUp, urlLogin) => {
     e.preventDefault();
@@ -54,46 +45,14 @@ const LoginInput = ({ isLoginMode }) => {
       .catch(setIsShowModal(true));
   };
 
-  const birthValidation = _inputBirth => {
-    let Birth_exp = /^(\(?\+?[0-9]*\)?)?[0-9_\- ]*$/;
-    let year = Number(_inputBirth.substring(0, 4));
-    let month = Number(_inputBirth.substring(4, 6));
-    let day = Number(_inputBirth.substring(6, 8));
-    let today = new Date();
-    let yearNow = today.getFullYear();
-
-    if (_inputBirth === undefined) {
-      return false;
-    } else if (_inputBirth.length !== 8 || !Birth_exp.test(_inputBirth)) {
-      return false;
-    } else if (1900 > year || year > yearNow) {
-      return false;
-    } else if (month < 1 || month > 12) {
-      return false;
-    } else if (day < 1 || day > 31) {
-      return false;
-    } else if (
-      (month === 4 || month === 6 || month === 9 || month === 11) &&
-      day === 31
-    ) {
-      return false;
-    } else if (month === 2) {
-      const leapYear = year % 4 === 0 && (year % 100 !== 0 || year % 400 === 0);
-      if (day > 29 || (day === 29 && !leapYear)) {
-        return false;
-      }
-    }
-    return true;
+  const togleModal = () => {
+    setIsShowModal(togleModal => !togleModal);
   };
 
-  const InputValueSignup =
-    inputValues.username.includes('@') &&
-    inputValues.password.length >= 4 &&
-    inputValues.password.length <= 10 &&
-    inputValues.name.length > 2 &&
-    birthValidation(inputValues.birth) &&
-    inputValues.contact.length >= 10 &&
-    inputValues.address.length >= 12;
+  const handleInput = event => {
+    const { name, value } = event.target;
+    setInputValues({ ...inputValues, [name]: value });
+  };
 
   const InputValueLogin =
     inputValues.username.includes('@') && inputValues.password.length >= 5;
@@ -145,8 +104,10 @@ const LoginInput = ({ isLoginMode }) => {
 
       <button
         className="loginSignupButton"
-        disabled={isLoginMode ? !InputValueSignup : !InputValueLogin}
-        onClick={loginSignUp}
+        disabled={
+          isLoginMode ? !InputValueSignup(inputValues) : !InputValueLogin
+        }
+        onClick={() => (isLoginMode ? alert('회원가입!') : loginSignUp())}
       >
         <strong>{isLoginMode ? '회원가입' : '로그인'}</strong>
       </button>
